# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type BrandProfileDensity =
  (typeof BrandProfileDensity)[keyof typeof BrandProfileDensity];

export const BrandProfileDensity = {
  spacious: "spacious",
  balanced: "balanced",
  dense: "dense",
} as const;

export interface BrandProfile {
  id: string;
  /** Hex color code e.g. */
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  headingFont: string;
  bodyFont: string;
  logoObjectPath?: string | null;
  density: BrandProfileDensity;
  updatedAt: string;
}

export type UpdateBrandProfileBodyDensity =
  (typeof UpdateBrandProfileBodyDensity)[keyof typeof UpdateBrandProfileBodyDensity];

export const UpdateBrandProfileBodyDensity = {
  spacious: "spacious",
  balanced: "balanced",
  dense: "dense",
} as const;

export interface UpdateBrandProfileBody {
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  headingFont?: string;
  bodyFont?: string;
  logoObjectPath?: string | null;
  density?: UpdateBrandProfileBodyDensity;
}

export type ProjectDensity =
  (typeof ProjectDensity)[keyof typeof ProjectDensity];

export const ProjectDensity = {
  spacious: "spacious",
  balanced: "balanced",
  dense: "dense",
} as const;

export interface Project {
  id: string;
  name: string;
  description: string;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  headingFont: string;
  bodyFont: string;
  logoObjectPath?: string | null;
  density: ProjectDensity;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyDensity =
  (typeof CreateProjectBodyDensity)[keyof typeof CreateProjectBodyDensity];

export const CreateProjectBodyDensity = {
  spacious: "spacious",
  balanced: "balanced",
  dense: "dense",
} as const;

export interface CreateProjectBody {
  name: string;
  description?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  headingFont?: string;
  bodyFont?: string;
  density?: CreateProjectBodyDensity;
}

export type UpdateProjectBodyDensity =
  (typeof UpdateProjectBodyDensity)[keyof typeof UpdateProjectBodyDensity];

export const UpdateProjectBodyDensity = {
  spacious: "spacious",
  balanced: "balanced",
  dense: "dense",
} as const;

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  headingFont?: string;
  bodyFont?: string;
  logoObjectPath?: string | null;
  density?: UpdateProjectBodyDensity;
}

export interface SlideOutline {
  /** 0-based slide index */
  slideIndex: number;
  /** Specific instructions for this slide */
  guidance: string;
}

export type SlideLayoutType =
  (typeof SlideLayoutType)[keyof typeof SlideLayoutType];

export const SlideLayoutType = {
  title: "title",
  section: "section",
  text: "text",
  columns: "columns",
  quote: "quote",
  metrics: "metrics",
} as const;

export interface SlideMetric {
  value: string;
  label: string;
}

export interface Slide {
  slideIndex: number;
  title: string;
  /** Markdown body content */
  body: string;
  layoutType: SlideLayoutType;
  speakerNotes: string;
  bulletPoints?: string[];
  metrics?: SlideMetric[];
  columnLeft?: string | null;
  columnRight?: string | null;
}

export type DeckNarrativeStructure =
  (typeof DeckNarrativeStructure)[keyof typeof DeckNarrativeStructure];

export const DeckNarrativeStructure = {
  "problem-solution": "problem-solution",
  consulting: "consulting",
  chronological: "chronological",
  "mece-pyramid": "mece-pyramid",
  "executive-summary": "executive-summary",
  custom: "custom",
} as const;

export interface Deck {
  id: string;
  title: string;
  brief: string;
  audience: string;
  narrativeStructure: DeckNarrativeStructure;
  slides: Slide[];
  projectId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DeckSummary {
  id: string;
  title: string;
  brief: string;
  slideCount: number;
  projectId?: string | null;
  createdAt: string;
}

export type GenerateDeckBodyNarrativeStructure =
  (typeof GenerateDeckBodyNarrativeStructure)[keyof typeof GenerateDeckBodyNarrativeStructure];

export const GenerateDeckBodyNarrativeStructure = {
  "problem-solution": "problem-solution",
  consulting: "consulting",
  chronological: "chronological",
  "mece-pyramid": "mece-pyramid",
  "executive-summary": "executive-summary",
  custom: "custom",
} as const;

export interface GenerateDeckBody {
  title: string;
  brief: string;
  audience: string;
  /**
   * @minimum 3
   * @maximum 30
   */
  slideCount: number;
  narrativeStructure: GenerateDeckBodyNarrativeStructure;
  /** Optional project to use for branding and corpus scoping */
  projectId?: string | null;
  /** Optional per-slide guidance for slide-by-slide mode */
  slideOutlines?: SlideOutline[];
}

export type UpdateSlideBodyLayoutType =
  (typeof UpdateSlideBodyLayoutType)[keyof typeof UpdateSlideBodyLayoutType];

export const UpdateSlideBodyLayoutType = {
  title: "title",
  section: "section",
  text: "text",
  columns: "columns",
  quote: "quote",
  metrics: "metrics",
} as const;

export interface UpdateSlideBody {
  title?: string;
  body?: string;
  bulletPoints?: string[];
  speakerNotes?: string;
  layoutType?: UpdateSlideBodyLayoutType;
  metrics?: SlideMetric[];
  columnLeft?: string | null;
  columnRight?: string | null;
}

export interface RegenerateSlideBody {
  /** Optional instruction for how to regenerate the slide */
  instruction?: string;
}

export interface DeckStats {
  totalDecks: number;
  totalCorpusDocuments: number;
  totalSlidesGenerated: number;
  totalChunks: number;
}

export type CorpusDocumentFileType =
  (typeof CorpusDocumentFileType)[keyof typeof CorpusDocumentFileType];

export const CorpusDocumentFileType = {
  pdf: "pdf",
  pptx: "pptx",
} as const;

export type CorpusDocumentStatus =
  (typeof CorpusDocumentStatus)[keyof typeof CorpusDocumentStatus];

export const CorpusDocumentStatus = {
  processing: "processing",
  ready: "ready",
  error: "error",
} as const;

export interface CorpusDocument {
  id: string;
  filename: string;
  fileType: CorpusDocumentFileType;
  chunkCount: number;
  status: CorpusDocumentStatus;
  projectId?: string | null;
  createdAt: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type UpdateProjectLogoBody = {
  objectPath: string;
};

export type ListCorpusDocumentsParams = {
  /**
   * Filter documents by project
   */
  projectId?: string;
};

export type UploadCorpusDocumentBody = {
  file: Blob;
  /** Optional project to associate the document with */
  projectId?: string;
};