# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DeckAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /decks/generate:
    post:
      operationId: generateDeck
      tags: [decks]
      summary: Generate a new deck from a brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDeckBody"
      responses:
        "201":
          description: Generated deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
...
  /decks/{id}:
    get:
      operationId: getDeck
      tags: [decks]
      summary: Get a specific deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "404":
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDeck
      tags: [decks]
      summary: Delete a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
  /decks/{id}/slides/{slideIndex}:
    put:
      operationId: updateSlide
      tags: [decks]
      summary: Update a slide inline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: slideIndex
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSlideBody"
      responses:
        "200":
          description: Updated deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "404":
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /decks/{id}/slides/{slideIndex}/regenerate:
    post:
      operationId: regenerateSlide
      tags: [decks]
      summary: Regenerate a single slide
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: slideIndex
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateSlideBody"
      responses:
        "200":
          description: Updated deck with regenerated slide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "404":
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /decks/{id}/export:
    get:
      operationId: exportDeck
      tags: [decks]
      summary: Export deck as PPTX
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PPTX binary file
          content:
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema:
                type: string
                format: binary
        "404":
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /corpus:
    get:
      operationId: listCorpusDocuments
      tags: [corpus]
      summary: List corpus documents
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: string
          description: Filter documents by project
      responses:
        "200":
          description: List of corpus documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorpusDocument"
  /corpus/upload:
    post:
      operationId: uploadCorpusDocument
      tags: [corpus]
      summary: Upload a PDF or PPTX to the corpus
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                projectId:
                  type: string
                  description: Optional project to associate the document with
              required:
                - file
      responses:
        "201":
          description: Uploaded corpus document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorpusDocument"
  /corpus/{id}:
    delete:
      operationId: deleteCorpusDocument
      tags: [corpus]
      summary: Delete a corpus document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted successfully
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SlideOutline:
      type: object
      properties:
        slideIndex:
          type: integer
          description: 0-based slide index
        guidance:
          type: string
          description: Specific instructions for this slide
      required:
        - slideIndex
        - guidance
    Slide:
      type: object
      properties:
        slideIndex:
          type: integer
        title:
          type: string
        body:
          type: string
          description: Markdown body content
        layoutType:
          type: string
          enum: [title, section, text, columns, quote, metrics]
        speakerNotes:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/SlideMetric"
        columnLeft:
          type: string
          nullable: true
        columnRight:
          type: string
          nullable: true
      required:
        - slideIndex
        - title
        - body
        - layoutType
        - speakerNotes
    SlideMetric:
      type: object
      properties:
        value:
          type: string
        label:
          type: string
      required:
        - value
        - label
    Deck:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        brief:
          type: string
        audience:
          type: string
        narrativeStructure:
          type: string
          enum: [problem-solution, consulting, chronological, mece-pyramid, executive-summary, custom]
        slides:
          type: array
          items:
            $ref: "#/components/schemas/Slide"
        projectId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - brief
        - audience
        - narrativeStructure
        - slides
        - createdAt
        - updatedAt
...
    GenerateDeckBody:
      type: object
      properties:
        title:
          type: string
        brief:
          type: string
        audience:
          type: string
        slideCount:
          type: integer
          minimum: 3
          maximum: 30
        narrativeStructure:
          type: string
          enum: [problem-solution, consulting, chronological, mece-pyramid, executive-summary, custom]
        projectId:
          type: string
          nullable: true
          description: Optional project to use for branding and corpus scoping
        slideOutlines:
          type: array
          items:
            $ref: "#/components/schemas/SlideOutline"
          description: Optional per-slide guidance for slide-by-slide mode
      required:
        - title
        - brief
        - audience
        - slideCount
        - narrativeStructure
    UpdateSlideBody:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        speakerNotes:
          type: string
        layoutType:
          type: string
          enum: [title, section, text, columns, quote, metrics]
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/SlideMetric"
        columnLeft:
          type: string
          nullable: true
        columnRight:
          type: string
          nullable: true
    RegenerateSlideBody:
      type: object
      properties:
        instruction:
          type: string
          description: Optional instruction for how to regenerate the slide
...
    CorpusDocument:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        fileType:
          type: string
          enum: [pdf, pptx]
        chunkCount:
          type: integer
        status:
          type: string
          enum: [processing, ready, error]
        projectId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - filename
        - fileType
        - chunkCount
        - status
        - createdAt
...