# lib/api-zod/src/generated/types/corpusDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CorpusDocumentFileType } from "./corpusDocumentFileType";
import type { CorpusDocumentStatus } from "./corpusDocumentStatus";

export interface CorpusDocument {
  id: string;
  filename: string;
  fileType: CorpusDocumentFileType;
  chunkCount: number;
  status: CorpusDocumentStatus;
  projectId?: string | null;
  createdAt: Date;
}