# lib/api-zod/src/generated/types/deck.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeckNarrativeStructure } from "./deckNarrativeStructure";
import type { Slide } from "./slide";

export interface Deck {
  id: string;
  title: string;
  brief: string;
  audience: string;
  narrativeStructure: DeckNarrativeStructure;
  slides: Slide[];
  projectId?: string | null;
  createdAt: Date;
  updatedAt: Date;
}