# lib/api-zod/src/generated/types/generateDeckBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateDeckBodyNarrativeStructure } from "./generateDeckBodyNarrativeStructure";
import type { SlideOutline } from "./slideOutline";

export interface GenerateDeckBody {
  title: string;
  brief: string;
  audience: string;
  /**
   * @minimum 3
   * @maximum 30
   */
  slideCount: number;
  narrativeStructure: GenerateDeckBodyNarrativeStructure;
  /** Optional project to use for branding and corpus scoping */
  projectId?: string | null;
  /** Optional per-slide guidance for slide-by-slide mode */
  slideOutlines?: SlideOutline[];
}