# lib/api-zod/src/generated/types/slideOutline.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SlideOutline {
  /** 0-based slide index */
  slideIndex: number;
  /** Specific instructions for this slide */
  guidance: string;
}